+++ src/main/java6/net/finmath/montecarlo/hybridassetinterestrate/HybridAssetLIBORModelMonteCarloSimulation.java
@@ -12,7 +12,7 @@ import net.finmath.exception.CalculationException;
 import net.finmath.marketdata.model.curves.DiscountCurveInterface;
 import net.finmath.montecarlo.BrownianMotionInterface;
 import net.finmath.montecarlo.assetderivativevaluation.AssetModelMonteCarloSimulationInterface;
-import net.finmath.montecarlo.interestrate.LIBORMarketModelInterface;
+import net.finmath.montecarlo.interestrate.LIBORModelInterface;
 import net.finmath.montecarlo.interestrate.LIBORModelMonteCarloSimulationInterface;
 import net.finmath.montecarlo.process.AbstractProcessInterface;
 import net.finmath.stochastic.RandomVariableInterface;
@@ -157,7 +157,7 @@ public class HybridAssetLIBORModelMonteCarloSimulation implements HybridAssetLIB
 		return liborSimulation.getBrownianMotion();
 	}
 
-	public LIBORMarketModelInterface getModel() {
+	public LIBORModelInterface getModel() {
 		return liborSimulation.getModel();
 	}