+++ src/test/java6/net/finmath/tests/montecarlo/interestrate/LIBORMarketModelValuationTest.java
@@ -292,7 +292,7 @@ public class LIBORMarketModelValuationTest {
 			double discountFactor	= getSwapAnnuity(liborMarketModel, new double[] { periodStart , periodEnd}) / periodLength;
 			int optionMaturityIndex = liborMarketModel.getTimeIndex(optionMaturity);
 			int liborIndex = liborMarketModel.getLiborPeriodIndex(periodStart);
-			double volatility = Math.sqrt(liborMarketModel.getModel().getIntegratedLIBORCovariance()[optionMaturityIndex][liborIndex][liborIndex]/optionMaturity);
+			double volatility = Math.sqrt(((LIBORMarketModelInterface)liborMarketModel.getModel()).getIntegratedLIBORCovariance()[optionMaturityIndex][liborIndex][liborIndex]/optionMaturity);
 			double valueAnalytic = net.finmath.functions.AnalyticFormulas.blackModelDgitialCapletValue(forward, volatility, periodLength, discountFactor, optionMaturity, strike);
 			System.out.print(formatterValue.format(valueAnalytic) + "          ");